+++ src/library/audio/sdl/sdlaudio.cpp
@@ -265,6 +265,54 @@ static int open_audio_device(const SDL_AudioSpec * desired, SDL_AudioSpec * obta
     return NULL;
 }
 
+/* Override */ int SDL_GetAudioDeviceSpec(int index, int iscapture, SDL_AudioSpec *spec)
+{
+    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
+    if (iscapture != 0)
+        return -1;
+    if (!spec)
+        return -1;    
+    if (index != 0)
+        return -1;
+
+    spec->freq = shared_config.audio_frequency;
+    switch (shared_config.audio_bitdepth) {
+        case 8:
+            spec->format = AUDIO_U8;
+            break;
+        case 16:
+            spec->format = AUDIO_S16LSB;
+            break;
+    }
+    spec->channels = shared_config.audio_channels;
+
+    return 0;
+}
+
+/* Override */ int SDL_GetDefaultAudioInfo(char **name, SDL_AudioSpec *spec, int iscapture)
+{
+    DEBUGLOGCALL(LCF_SDL | LCF_SOUND);
+    if (iscapture != 0)
+        return -1;
+    if (!spec)
+        return -1;
+    if (name)
+        *name = nullptr;
+
+    spec->freq = shared_config.audio_frequency;
+    switch (shared_config.audio_bitdepth) {
+        case 8:
+            spec->format = AUDIO_U8;
+            break;
+        case 16:
+            spec->format = AUDIO_S16LSB;
+            break;
+    }
+    spec->channels = shared_config.audio_channels;
+
+    return 0;
+}
+
 /* Override */ SDL_AudioDeviceID SDL_OpenAudioDevice(const char *device,
                    int iscapture, const SDL_AudioSpec *desired,
                    SDL_AudioSpec *obtained, int allowed_changes)